liteconvert/requirements.txt
PyQt5>=5.15
Pillow>=10.0
pillow-heif>=0.16
img2pdf>=0.5
PyMuPDF>=1.24
pytest>=7.0

